backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
pydantic>=2.6.0
pydantic-settings>=2.1.0
requests>=2.31.0
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
email-validator>=2.1.0